Put + Validation Business Rules (Exception + Validation classes + injection in Service - business domain)

+++ src/main/java/com/mike/springjpalibrary/service/AuthorService.java
@@ -1,27 +1,39 @@
 package com.mike.springjpalibrary.service;
 
+import com.mike.springjpalibrary.exceptions.OperationNotAllowed;
 import com.mike.springjpalibrary.model.Author;
 import com.mike.springjpalibrary.repository.AuthorRepository;
+import com.mike.springjpalibrary.repository.BookRepository;
+import com.mike.springjpalibrary.validator.AuthorValidator;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
+//@RequiredArgsConstructor(onConstructor_ = @Autowired)
 @Service
 public class AuthorService implements IService<Author> {
 
-    private AuthorRepository authorRepository;
+    private final AuthorRepository authorRepository;
+    private final AuthorValidator authorValidator;
+    private final BookRepository bookRepository;
 
     @Autowired
-    public AuthorService(AuthorRepository authorRepository) { // bean gerenciado (repository)
+    public AuthorService(AuthorRepository authorRepository, AuthorValidator authorValidator, BookRepository bookRepository) { // bean gerenciado (repository)
         this.authorRepository = authorRepository;
+        this.authorValidator = authorValidator;
+        this.bookRepository = bookRepository;
     }
 
     @Override
     public Author save(Author author){
-            return authorRepository.save(author);
+        System.out.println("Saving Author");
+        authorValidator.validar(author);
+        return authorRepository.save(author);
     }
 
     @Override
@@ -36,30 +48,53 @@ public class AuthorService implements IService<Author> {
 
     @Override
     public void delete(Author author) {
+        if (authorHasBook(author)) {
+            throw new OperationNotAllowed("Author has books registerd");
+        }
         authorRepository.delete(author);
     }
 
     @Override
-    public Author update(Author author) {
-        return null;
+    public void update(Author author) {
+        if (author.getId() == null) {
+            throw new IllegalArgumentException("Author not registered");
+        }
+        authorValidator.validar(author);
+        authorRepository.save(author);
     }
 
     @Override
-    public List<Author> findByNameAndNationality(String name, String nationality) {
-        if(name == null && nationality == null) {
+    public List<Author> findByNameAndBirthDateAndNationality(String name, LocalDate birthDate, String nationality) {
+        if(name == null && nationality == null && birthDate == null) {
             return authorRepository.findAll();
         }
 
-        if (name == null) {
-            return authorRepository.findByNationality(nationality);
-        }
-
-        if(nationality == null) {
+        if (name != null && nationality == null && birthDate == null) {
             return authorRepository.findByNome(name);
         }
 
-        return authorRepository.findByNomeAndNationality(name, nationality);
+        if (name == null && nationality != null && birthDate == null) {
+            return authorRepository.findByNationality(nationality);
+        }
+        if (name == null && nationality == null) {
+            return authorRepository.findByBirthDate(birthDate);
+        }
+
+        if (name == null) {
+            return authorRepository.findByNationalityAndBirthDate(nationality, birthDate);
+        }
+        if (birthDate == null) {
+            return authorRepository.findByNomeAndNationality(name, nationality);
+        }
+        if (nationality == null) {
+            return authorRepository.findByNomeAndBirthDate(name, birthDate);
+        }
+
+        return authorRepository.findByNomeAndBirthDateAndNationality(name, birthDate, nationality);
     }
 
+    public boolean authorHasBook(Author author) {
+        return bookRepository.existsByAuthor(author);
+    }
 
 }
